Reorganise (simplify) files before starting project
