+++ app/javascript/scripts/component_activated_menu.js
