+++ app/javascript/scripts/editable_content.js
@@ -1,0 +1,1 @@
+console.log("Editable Content script loaded");
