+++ app/javascript/styles/_component_activated_menu.scss
